Fix matching engine logic
This commit fixes the logic in the matching engine (mostly contained
within `Book._match()`, in `pyobsim/book.py`) to iterate through the
price list for the side being matched against, regardless of any other
conditions (such as whether the order's price is in the price list, the
type of the side being matched against, etc.). This fixes several bugs
and also greatly simplifies the matching engine's code.